segmentation/requirements.txt
opencv-python~=4.10.0.84
opencv-python-headless~=4.10.0.84
albumentations~=1.4.11
pandas~=2.2.2
torch~=2.3.1
wandb~=0.17.4
numpy~=1.26.4
scipy~=1.14.0
PyYAML~=6.0.1
scikit-learn~=1.5.1
tqdm~=4.66.4
pillow~=10.4.0
timm==1.0.7
jupyterlab~=4.2.3
tabulate==0.9.0
transformers==4.51.3
open_clip_torch==2.32.0
tables==3.10.2
faiss-cpu==1.10.0
torchmetrics==1.7.1
pyarrow==20.0.0
git+https://github.com/facebookresearch/segment-anything.git

# URL to the custom wheel file
https://github.com/BohemianVRA/FGVC/releases/download/v1.5.4-dev/fgvc-1.5.4.dev0-py3-none-any.whl